feat: add Jest configuration and initial test suite for song statistics
