+++ shared/package.json
@@ -1,30 +1,35 @@
 {
-  "name": "noteblockworld-shared",
-  "version": "0.1.0",
-  "description": "Shared packages for noteblockworld",
-  "keywords": [],
-  "author": "",
-  "license": "UNLICENSED",
-  "scripts": {
-    "lint": "eslint */**/*.ts --fix"
-  },
-  "dependencies": {
-    "@encode42/nbs.js": "^5.0.0",
-    "@napi-rs/canvas": "^0.1.53",
-    "@nestjs/swagger": "^7.1.1",
-    "@timohausmann/quadtree-ts": "^2.2.2",
-    "@types/unidecode": "^0.1.3",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
-    "express": "^4.18.2",
-    "jszip": "^3.10.1",
-    "tailwindcss": "3.4.1",
-    "unidecode": "^1.1.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/multer": "^1.4.11",
-    "@types/node": "^20.3.1",
-    "typescript": "^5.1.3"
-  }
+    "name": "noteblockworld-shared",
+    "version": "0.1.0",
+    "description": "Shared packages for noteblockworld",
+    "keywords": [],
+    "author": "",
+    "license": "UNLICENSED",
+    "scripts": {
+        "lint": "eslint */**/*.ts --fix",
+        "test": "jest",
+        "test:watch": "jest --watch"
+    },
+    "dependencies": {
+        "@encode42/nbs.js": "^5.0.0",
+        "@napi-rs/canvas": "^0.1.53",
+        "@nestjs/swagger": "^7.1.1",
+        "@timohausmann/quadtree-ts": "^2.2.2",
+        "@types/unidecode": "^0.1.3",
+        "class-transformer": "^0.5.1",
+        "class-validator": "^0.14.0",
+        "express": "^4.18.2",
+        "jszip": "^3.10.1",
+        "tailwindcss": "3.4.1",
+        "unidecode": "^1.1.0"
+    },
+    "devDependencies": {
+        "@types/express": "^4.17.17",
+        "@types/jest": "^29.5.2",
+        "@types/multer": "^1.4.11",
+        "@types/node": "^20.3.1",
+        "jest": "^29.5.0",
+        "ts-jest": "^29.1.0",
+        "typescript": "^5.1.3"
+    }
 }
